+++ src/ts/utils.ts
@@ -3,12 +3,12 @@
  */
 
 //  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
-//   _   _ _   _ _ _ _   _           
-//  | | | | |_(_) (_) |_(_) ___  ___ 
+//   _   _ _   _ _ _ _   _
+//  | | | | |_(_) (_) |_(_) ___  ___
 //  | | | | __| | | | __| |/ _ \/ __|
 //  | |_| | |_| | | | |_| |  __/\__ \
 //   \___/ \__|_|_|_|\__|_|\___||___/
-//                                   
+//
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 /**
@@ -38,6 +38,19 @@ export const bring: Function = (url: string, data: any): Promise<any> => {
  * @returns
  */
 export const eId: Function = (id: string): HTMLElement | undefined => {
-  const element = document.getElementById(id) 
+  const element = document.getElementById(id)
   return element ? element : undefined
+}
+
+/**
+ * 
+ * @param defaultParameters 
+ * @param inputParameters 
+ * @returns 
+ */
+export function applyDefaultParameters<Type, iType>(
+  defaultParameters: Type,
+  inputParameters: iType
+): Type {
+  return { ...defaultParameters, ...inputParameters }
 }